dockerfile-practice/python-interview/requirements.txt
flask
gunicorn
python-dotenv
psycopg2-binary  # If you use Postgres

